+++ SleepHunter/IO/Process/MemoryVariableExtender.cs
@@ -20,7 +20,6 @@ namespace SleepHunter.IO.Process
             reader.BaseStream.Position = address;
             var reference = reader.ReadUInt32();
 
-            Debug.WriteLine($"Dereferenced [{address.ToString("X")}] -> {reference.ToString("X")}");
             return reference;
         }