+++ SleepHunter/Services/IReleaseService.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Threading.Tasks;
+using SleepHunter.Models;
+
+namespace SleepHunter.Services
+{
+    public interface IReleaseService
+    {
+        Uri GetLatestReleaseNotesUri();
+
+        Task<ReleaseVersion> GetLatestReleaseVersionAsync();
+
+        Task<ReleaseAsset> GetLatestReleaseAsync();
+
+        Task<string> DownloadLatestReleaseAsync(Uri downloadUri, IProgress<long> progress = null);
+    }
+}