+++ SleepHunter/Views/SpellTargetWindow.xaml.cs
@@ -3,7 +3,6 @@ using System.ComponentModel;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
-using System.Windows.Media;
 using System.Windows.Media.Animation;
 using System.Windows.Threading;