+++ CHANGELOG.md
@@ -4,10 +4,24 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.0.2] - Unreleased
+## [4.1.0] - 2023-04-10
 
 ### Added
 
+- New `Updates` section in `Settings` window allowing version checking
+- New `Auto-Update` window for downloading updates and launching the updater
+- New `SleepHunter.Updater` child project for applying auto-updates
+- New scanline overlay for displaying modals
+- Check for new updates on startup (if enabled)
+
+### Changed
+
+- Accessor key for `About` is now `Alt+B` instead of `Alt+A` (conflicted with `All Macros`)
+
+### Removed
+
+- Dead code for MVVM (use `CommunityToolkit.Mvvm` instead)
+- Old `Debug.WriteLine` calls (only used in Debug mode)
 
 ## [4.0.1] - 2023-04-09